requirements.txt
# --- Sift AI Dependencies ---
# Core Network & API
requests>=2.32.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.1

# AI Providers (2026 SDKs)
openai>=1.60.0
google-genai>=0.5.0

# Document Processing (Soft Dependencies)
beautifulsoup4>=4.13.0
PyMuPDF>=1.25.0
python-docx>=1.2.0
odfpy>=1.4.1
striprtf>=0.0.26

# Dynamic Web Loading
playwright>=1.49.0

# GUI Theme (Sun Valley)
sv-ttk>=2.6.0